Remove the mca and pcr analysis from 2017-09-28
This folder mainly concentrates on the results using the NgAqua dataset, and
the results for homo2d are also included because  that is how the code was
developed. The scripts were generalized first using the original analysis of
the 2d dataset. This directory should be used for the modern versions of the
code.