WARNING serious error with linear regression report
I tried to reproduce the results of my own report to Chris, and I was getting
wildly different results than before. Namely the linear regression does not
return the same output whenever it is called.

Upon a rerun of the full analysis pipeline the linear regerssion are now wildly
unstable and give nonsense answers. This implies that we will need to use some
sort of regularization in the future.