use auto-encoders to denoise the forcing dataset
I demonstrate in this notebook, that the noise in the forcing dataset
results from the noisiness in the extremely stable region above the
tropopause. This strong stability tends to amplify small amount of noise
in W.

Using auto-encoders to compute the vertical advection substantially
reduces this noiseness. Leaving out the upper 10 grid points is
essential.