Use dict as the internal data structure for pytorch modules
I was previously using a concatenated data array as the basic input to pytorch,
but this approach is not very flexible, and does not provide an easy or robust
way to access individual physical variables. Being able to do so is important
for including things like surface fluxes or compute the precipitation for the
objective function.

this commit was a really extensive refactor that required very carefully using a
regression test to ensure the output did not change.

Move trapezoid rule out of ForcedStepper module

move logic from EulerStepper to the ForcedStepper module

do not pass forcings to the rhs operator

pass scaler to RHS

using nn.Sequential is more brital

RHS class expects a dict for inputs

RHS now outputs a dict as well as accepting a dict

use dicts for the time stepping

fix the loss function closure

the full training pipeline works again. Yay. I still have a lot of work to do
for this refactor though.

replace hardcoded array/dict conversions with functions

also output the loss

This is useful for regression testing purposes

change forced stepper input to dict

use a dict for the forcings also

Output of forced stepper is a dict

loss function takes dict input

compute loss on key by key basis

allow the wraps method to work with dictionary input and output