use module level variable to specify variable names
This commit means that I don't have to change both _data_to_torch_dataset and
the closure within train_multistep_objective just to add a variable to the data
dictionary.

It might be better in the future to make prognostic_variables and
forcing_variables attributes of some object. _data_to_torch_dataset and the data
pulling functionality could be methods of this object