Document how to match solar params between CAM and SAM
I used the notebook to help backwards engineer how the lat/lon and time
of day were used in the EQX run.

I took very specific notes in the docstring of lib/create_iopfile.py on
how to specify the namelist entries for SAM and the variables in the
IOPfile to handle the diurnal cycle properly.