Add adiabatic term to divT3D
This makes the CAM model run correctly on the NGAQaua inputs. this was the most
ridiculous amount of work ever. I am really shocked by how incomprehnsible the
CESM source code is. It is littered with gotos, completely opaque variables
names like qminus, q3, qfcst , divT etc...

One problem, is that if vertdivq and divq are both present, SCAM does add them
together into a variable called `divq3d`. However, it will only use `divq3d` to
force the humidity if `divq` is not present. This is basically a catch 22. Why
does the model not use a the vertdivq if I give it!! On top of that, I think the
model has a bug where it does not do vertical advection of the humidity if the
full 3d forcings are not used.

The ultimate solution I made is in lib.scam.create_iopfile.compute_tendencies. I
basically need to

1. Put total horz and vertical humidity advection terms in a variable called
`Q_dten`

2. Put the total advective tendency of _dry static energy_ in divT3d. Even
though this variable is called divT3d, it actually needs to include the g/cp w
term.

I computed the transport tendencies using the advective formultion u.grad(q),
where approximated grad(q) using centered differences.