Turn off negative moisture thresholding
THIS IS AN IMPORTANT COMMIT

For some reason, the precipitation I compute is drastically wrong when I
ensure the humidity is positive. I only found this out after significant
trial and error and by comparing to older working commits.

This solution is unsatisfying because the small amount of humidity in
the upper atmoshpere should not effect the results so strongly.