Delete unused classes and move torch subpackage
In preparation for publication I am removing a lot of the unnessary old cruft
from this project. This includes the sklearn classes I wrote for MCA. The
variational auto-encoder codes I wrote, and some other useless scripts.

It is possible that some of this stuff should be moved to my general purpose
repository.