Improve numpy-torch interface
The old interface (step_model) was very brittle because it tried to do too much.
It handled unit conversion, and had many hardcoded names and strings. The new
interface (step_with_numpy_inputs) is much cleaner, because it places the
responsibility of the units being correct on the calling function. It only
reshapes the numpy arrays and pushes them to torch.