Add PyTorch layers for dictionary input/output
I add two modules. One which does a linear transformation of dictionary
input, and another which does a linear transform with dictionary output.

These classes some benefits

1. Each variable has its own linear transform object, which makes it
easy to access the weights of a specific physical field. this could be
used, for example, to put a penalty on the weights for the FQT and FSLI
fields.
2. No concatenating is done before input, which consumes extra memory.