Run SAM with NN and hyperdiffusion
I hoped the model would work with hyper diffusion, but the results I am
getting are not improved by much.

The strangest thing about these initial condition is that the net
vertical motion in the tropics quickly vanishes. This happens both with
and w/o the neural network. Should the dry dynamics respond this way?

Maybe, I should test the response the model using a Gill-type heating
profile. Is the model working as expected or is something wrong with
rotation. Are there metric terms?