Output and analyze FQTNN and FSLINN
This large commit modifies he neural network code to save the
parametrized sources of humidity and temperature to disk. This is
necessary to understand why the model is failing in 3D simulations.

There are also many changes to the environment.yml, Docker image and
scripts for running the neural network inside SAM.

Resolves #8