Apply conservation to tendencies
Previously, I was updating the modified prognostic variables using the
conservation principals. Because the size of the time-step implicitly
effects this updated state, this led to problems when applying the the
neural network with a shorter time step than it was trained with.

The energy and moisture conservation laws can constrain the *tendencies*
directly.

Resolves #10