Use sacred for logging experiments
Before I was using a custom mongodb logger and comet ml. These required special
configuration and were not as robust as the experiment loggers provided by
sacred.

The configuration used to be scattered between a yaml file and command line
arguments. Sacred allows a more consistent approach to configuring experiments.

I also changed the routine for generating directories to save the data in. The
folders for models now look like "1-{tag}".

Resolves #13