Refactor torch model code
The MLP class is well on its way to becoming a "god" class, so I had to refactor
out the important bits.

The MLP function is split into two classes. 1) ForcedStepper which handles the
time stepping, and knows which variables are prognostic, etc. and 2)
ApparentSource which contains the actual neural network code.