Improve the multiple time step implementation
Previously, I had implemented this as a ForcedStepper class, but that
class was making errors that were hard to trace. Now, the multiple time
stepper is a function rather than a torch module. Implementing a
polymorphic object (Batch) for grabbing fields and time steps from a mini-batch
made the code much cleaner.

A quickstart guide documents how to use this code.

Finally, this new code is used to study the spin-up error made by a
multiple time stepper. The main conclusion is that spin-up matters more
in moisture than temperature and is most important in the upper
atmosphere.