Penalize equilibration for one step error
This commit potentially resolves a problem with bias due to the multiple
time step prediction.  Instead of computing the multiple step error, I
add a penalty term which penalizes the growth of systems under constant
forcing. This allows a stable neural network to be trained with only one
time step.

I also converted the training function to use pytorch-ignite for running
through the batches.