Add linearized analysis functions
I hypothesize that blow-up in SAM is connected to linearly unstable
modes of the neural network. This commit introduces code for computing
the jacobian and it's eigenvalues in an autodiff friendly way. Also,
the weak temperature gradient in the tropics significantly alters what
the unstable modes will look like, so I include code that penalizes
unstable WTG modes.

These changs required some additions to the Batch class in the
timestepper module.