Experiment debias layer (#47)
This model debiases the outputs of a Neural network using a LASSO technique. This should remove any bias when evaluated semi-prognostically, but it cannot control coupled biases.

* Add precipitation pdf plots

* Add jupyter notebook with LASSO based debiasing

I fit a different LASSO model in each height and y location, and then
use this to debiase the data. This gives seems to give nice results.

* Add sam_interface debugging file to dvc

* Add 'test' target to makefile

this speeds up the test running procedure.

* Add torch model wrapper classes

* add NumpyWrapper and XarrayWrapper classes for torch models. NumpyWrapper
allows passing dicts of numpy arrays to a torch model.
* Use these wrapper classes in sam_interface.py. Previously, this module relied
on call_model_with_numpy. Now, it can load any object with the right calling
interface.

* Add LassoDebiasedModel which is trained with a dataarray

* Add numpy interface to LassoDebiasedModel

* Automate debiasing with snakemake