Feature/turn off stability (#72)
* Configure weight of instability penalty

Previously we assumed that loss = l_consistency + l_stability, and could
not turn off the penalty.

I made the configuration of the stepper more versatile and added an
option "alpha" for so that loss = l_consit + alpha * l_stab

* Run simulation with no instability penalty

The simulation is located at:

data/runs/model280-epoch5.debiased/

This simulation does not blow up immediately but does develop
grid-scale storms. This could occur because of the debiasing procedure
though.