Couple NN to linear wave dynamics (#71)
Coupling the NN to linear wave dynamics (Kuang 2018) can reveal the types of instability produced by the NN model when coupled to a GCM.

This commit introduces a `Wave2D` for the dry waves which is coupled to the LRF of a neural network using the `WaveCoupler` class. The wave spectra of the unstable and stable NNs are compared using this method.