Run SAM at 40 km resolution (#73)
By running SAM at 40 km resolution, but with the 160km-trained NN, we hope to see if this scheme produces more precipitation. It does not, and the simulation looks very similar to a control.

- compile and run SAM with two different resolutions: 160km and 40km
- works in docker and on olympus

Contains the following squashed commits:

* Reprocess data using SAM

* add a new parameters file
* use the master run_sam_ic_nn.py script and delete old process_ngaqua.py
  script. This simplifies the SAM running pipeline

* Use a xarray for zonal+time_mean

Xarray is faster than NCO because it can use multithreading.

* Implement flexible initial conditions

* Specify grid size at compile time using macros

Compiling SAM to run at different resolutions required manually editing the
domain.f90 file, which is not good.

This commit improves this by using the cpp preprocessor, which allows
specifying the grid size on compile time. For now, the resolution is specified
in the ./Build file with

setenv NX ...
setenv NY ...
setenv NZ ...
setenv NSUBX ...
setenv NSUBY ...

the built SAM executable is availabe in the folder named like:

OBJ/512x256x34/

* Add Peter's alias to environment

* Read netCDF file in parallel

* Specify path to RUNDATA in create_case.py

Previously this path was inferred from the sam source root. However, the new
resolution specific compiling required specificng the sam root as
OBJ/<resolution spec>. Because RUNDATA is not located at this path, we need to
provide a new option to specify its location.

* Make hyperdiffusion work with mpi

Previously, the halo communication was disabled.

This commit performs communication with periodic(4) and adds flags for
disabling hyper-diffusion altogether.

* Correct boundary handling for NN+GCM+NPI

The meridional boundary region where the NN is not applied is computed from the
edge of each processors subdomain. For parallel jobs, this should only be done
in the northernmost or southernmost subdomains.

I also removed some unnessary logging information and time the different stages
of the python code.

* Compile different resolutions in docker image

* Update sam integration test to work with resolution

The commit changing the resolution and location of the paths broke
this test.

* Add configurations for sam high resolutions

* Add "enter" target to makefile to start docker

* Compare the 40 km and 160 km simulations

Unfortunately, the 40 km run does not rain more/less in the tropics than the
160 km one.