Train/test model with blurred data (#74)
To simulate the effect of coarse-resolution data, we can test/train the NN on blurred training data.

Changes needed:

* Add script for blurring the data

* Refactor and improve SAM-based pre-processing scripts.

The pre-processing pipeline was too confusing, and was scatter across several snakemake rules. Now these are combined into one script: `uwnet/data/preprocess.py.`

* Automate NN training and SAM simulation with snakemake

These steps had to be executed manually because Sacred generated the folder names automatically. Now the model training and SAM runs are named based on the filename of the json file used to train them.

* Improve training messages

Previously it was hard debugging errors with the input data. Also use the agg backend for plots, so that the training does not die on olympus.