Restore and improve the nudging performance (#11)
* Restore nudging capability

This was broken by some changes to the training data (getting rid of the 'step')
dimension.

I also refactored the create_case script to be more clean.

* Profile the python caller internally

Add initialize and finalize hooks to python caller. The intialize function
starts a profiler, and the finalize one stops it.

* Install netCDF4 in the docker environment explicitly

This resolves #9, but I am not sure why.

* Fix name error for Profile

* Improve the performance of the nudger

The conversion from xarray to numpy and reloading the same time points
repeatedly were slow. This commit still uses xarray to load the target data, but
it uses the lru_cache decorator to cache these results. The function also
outputs numpy data, so the xarray to numpy conversion is also cached, and all
mathematical operations occur on raw numpy arrays.

Resolves #10